src/requirements.txt
Flask
flask-injector
flask-session
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
msal
python-dateutil